+++ controller/src/extension.ts
@@ -514,3 +514,5 @@ export function activate(context: vscode.ExtensionContext) {
 function cloneSelectedRows(context: Record<string, any>) {
   throw new Error("Function not implemented.");
 }
+
+